+++ server/src/weaverbird/backends/pypika_translator/translators/base.py
@@ -3,7 +3,7 @@ from collections.abc import Callable, Mapping, Sequence
 from dataclasses import dataclass
 from datetime import UTC, datetime
 from functools import cache
-from typing import TYPE_CHECKING, Any, Literal, TypeVar, Union, cast, get_args
+from typing import TYPE_CHECKING, Any, TypeVar, Union, cast, get_args
 
 from dateutil import parser as dateutil_parser
 from pypika import (
@@ -19,7 +19,7 @@ from pypika import (
     analytics,
     functions,
 )
-from pypika.enums import Comparator, Dialects, JoinType
+from pypika.enums import Comparator, Dialects
 from pypika.functions import Extract, ToDate
 from pypika.queries import QueryBuilder, Selectable
 from pypika.terms import (
@@ -46,11 +46,10 @@ from weaverbird.pipeline.conditions import (
     NullCondition,
 )
 from weaverbird.pipeline.dates import RelativeDate
-from weaverbird.pipeline.pipeline import Pipeline
 from weaverbird.pipeline.steps import ConvertStep, DomainStep, FilterStep, TopStep
 from weaverbird.pipeline.steps.date_extract import DATE_INFO
 from weaverbird.pipeline.steps.duration import DURATIONS_IN_SECOND
-from weaverbird.pipeline.steps.utils.combination import PipelineOrDomainNameOrReference, Reference
+from weaverbird.pipeline.steps.utils.combination import Reference
 
 from .exceptions import ForbiddenSQLStep, UnknownTableColumns
 
@@ -62,7 +61,6 @@ if TYPE_CHECKING:
     from weaverbird.pipeline.steps import (
         AbsoluteValueStep,
         AggregateStep,
-        AppendStep,
         ArgmaxStep,
         ArgminStep,
         CompareTextStep,
@@ -78,7 +76,6 @@ if TYPE_CHECKING:
         FormulaStep,
         FromdateStep,
         IfthenelseStep,
-        JoinStep,
         LowercaseStep,
         PercentageStep,
         RankStep,
@@ -509,67 +506,6 @@ class SQLTranslator(ABC):
                 merged_query.orderby(*step.on) if step.on else merged_query,
                 selected_col_names,
             )
-
-    @staticmethod
-    def _pipeline_or_domain_name_or_reference_to_pipeline(
-        pipeline: "PipelineOrDomainNameOrReference", message: str | None = None
-    ) -> list["PipelineStep"]:
-        try:
-            return Pipeline(steps=pipeline).steps
-        except Exception as exc:
-            message = message or f"could not convert {pipeline} to pipeline: {exc}"
-            raise NotImplementedError(message) from exc
-
-    def append(
-        self: Self,
-        *,
-        builder: "QueryBuilder",
-        prev_step_table: str,
-        columns: list[str],
-        step: "AppendStep",
-    ) -> StepContext:
-        pipelines = [self._pipeline_or_domain_name_or_reference_to_pipeline(pipeline) for pipeline in step.pipelines]
-        tables: list[str] = []
-        column_lists: list[list[str]] = []
-        for pipeline in pipelines:
-            pipeline_ctx = self.__class__(
-                tables_columns=self._tables_columns,
-                db_schema=self._db_schema_name,
-                known_instances=self._known_instances,
-            ).get_query_builder(steps=pipeline, query_builder=builder)
-            tables.append(pipeline_ctx.table_name)
-            column_lists.append(pipeline_ctx.columns)
-            builder = pipeline_ctx.builder
-
-        # By default, UNION ALL will append columns by index in SQL, and only append as much columns
-        # as available in the first SELECT statement. Also, the columns are appended in the order of
-        # the select statement, which is not what we want either (a merge by name is probably what's
-        # expected by the user)
-        columns_to_add = []
-        for column_list in column_lists:
-            for col in column_list:
-                if col not in columns and col not in columns_to_add:
-                    columns_to_add.append(col)
-
-        all_columns = columns + columns_to_add
-
-        query = self.QUERY_CLS.from_(prev_step_table).select(
-            *columns, *(LiteralValue("NULL").as_(col) for col in columns_to_add)
-        )
-        for table, column_list in zip(tables, column_lists, strict=True):
-            query = query.union_all(
-                self.QUERY_CLS.from_(table).select(
-                    *(
-                        # Selecting either the column from the dataset if it is available, or NULL
-                        # AS "col" otherwise. We iterate over all_columns rather than column_list in
-                        # order to have a merge by name
-                        col if col in column_list else LiteralValue("NULL").as_(col)
-                        for col in all_columns
-                    )
-                )
-            )
-
-        return StepContext(query.orderby(*all_columns), all_columns, builder)
 
     def argmax(
         self: Self,
@@ -1255,67 +1191,8 @@ class SQLTranslator(ABC):
         return StepContext(query, columns + [step.new_column])
 
     @staticmethod
-    def _get_join_type(join_type: Literal["left", "inner", "left outer"]) -> JoinType:
-        match join_type:
-            case "left":
-                return JoinType.left
-            case "left outer":
-                return JoinType.left_outer
-            case "inner":
-                return JoinType.inner
-
-    @staticmethod
     def _field_list_to_name_list(fields: list[Field]) -> list[str]:
         return [field.alias or field.name for field in fields]
-
-    def join(
-        self: Self,
-        *,
-        builder: "QueryBuilder",
-        prev_step_table: str,
-        columns: list[str],
-        step: "JoinStep",
-    ) -> StepContext:
-        steps = self._pipeline_or_domain_name_or_reference_to_pipeline(step.right_pipeline)
-
-        right_builder_ctx = self.__class__(
-            tables_columns=self._tables_columns,
-            db_schema=self._db_schema_name,
-            known_instances=self._known_instances,
-        ).get_query_builder(steps=steps, query_builder=builder)
-        left_table = Table(prev_step_table)
-        right_table = Table(right_builder_ctx.table_name)
-
-        left_cols = [Field(col, table=left_table) for col in columns]
-        # Simply checking if a column is in left columns and right columns is not enough to
-        # determine if it needs to be aliased. In case of nested joins where a column is used in
-        # three or more datasets (not unlikely with columns like "id" or "name"), the aliased column
-        # may need to be suffixed several times. The join/inner_pypika_nested.yaml fixture shows an
-        # example of this kind of situation
-        right_cols: list[Field] = []
-        for col in right_builder_ctx.columns:
-            if col not in (all_cols := self._field_list_to_name_list(left_cols + right_cols)):
-                right_cols.append(Field(col, table=right_table))
-                continue
-            alias = col
-            while (alias := f"{alias}_right") in all_cols:
-                pass
-            right_cols.append(Field(col, table=right_table, alias=alias))
-
-        query = (
-            self.QUERY_CLS.from_(left_table)
-            .select(*left_cols, *right_cols)
-            .join(right_table, self._get_join_type(step.type))
-            .on(Criterion.all(Field(f[0], table=left_table) == Field(f[1], table=right_table) for f in step.on))
-            # Order of results is not consistent depending on the SQL Engine (inconsistencies
-            # observed with Athena and BigQuery).
-            .orderby(*(c[0] for c in step.on))
-        )
-        return StepContext(
-            query,
-            [c.name for c in left_cols] + [c.alias or c.name for c in right_cols],
-            right_builder_ctx.builder,
-        )
 
     def lowercase(
         self: Self,